Moved Standard Page base class to common folder
