+++ di_website/home/models.py
@@ -13,6 +13,7 @@ from wagtail.snippets.models import register_snippet
 
 from modelcluster.fields import ParentalKey
 from modelcluster.models import ClusterableModel
+from common.base import StandardPage
 
 
 class AbstractLink(models.Model):
@@ -136,40 +137,6 @@ class FooterText(models.Model):
 
     class Meta:
         verbose_name_plural = 'Footer Text'
-
-
-class StandardPage(Page):
-    class Meta:
-        abstract = True
-
-    hero_image = models.ForeignKey(
-        'wagtailimages.Image',
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name='+',
-        help_text='Hero Image'
-    )
-    hero_text = RichTextField(
-        null=True,
-        blank=True,
-        help_text='A description of the page content'
-    )
-    hero_link = models.ForeignKey(
-        'wagtailcore.Page',
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name='+',
-        verbose_name='Hero link',
-        help_text='Choose a page to link to for the Call to Action'
-    )
-    hero_link_caption = models.CharField(
-        max_length=255,
-        blank=True,
-        help_text='Text to display on the link button'
-    )
-
 
 class HomePage(StandardPage):
     content_panels = StandardPage.content_panels + [
